# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mateen/admin/sources:
    get:
      operationId: getSourceReviews
      responses:
        "200":
          description: Versioned Nawawi source review queue
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/SourceVersion" } }
    post:
      operationId: createSourceVersion
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SourceVersionInput" }
      responses:
        "201":
          description: New unapproved immutable version
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceVersion" }
  /mateen/admin/sources/{versionId}/decision:
    post:
      operationId: decideSourceVersion
      parameters:
        - { in: path, name: versionId, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SourceDecisionInput" }
      responses:
        "200":
          description: Scientific and rights decision on exact version
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceVersion" }
...
  /mateen/texts/{textId}:
    get:
      operationId: getStudyText
      parameters:
        - in: path
          name: textId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Available text and source records
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudyText" }
  /mateen/profile:
    get:
      operationId: getProfile
      responses:
        "200":
          description: Current authenticated profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
    put:
      operationId: saveProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfileInput" }
      responses:
        "200":
          description: Saved profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
...
  /mateen/progress/{textId}:
    put:
      operationId: saveProgress
      parameters:
        - in: path
          name: textId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProgressInput" }
      responses:
        "200":
          description: Saved progress, not an assessment result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudyProgress" }
...
components:
  schemas:
...
    Hadith:
      type: object
      required: [id, number, title, text, sourceUrl, sourcePage]
      properties:
        id: { type: integer }
        number: { type: integer }
        title: { type: string }
        text: { type: string }
        sourceUrl: { type: string }
        sourcePage: { type: integer }
        sourceVersionId: { type: string }
        reviewStatus: { type: string, enum: [pending_review, approved, rejected, withdrawn] }
        viewerPage: { type: integer }
    StudyText:
      type: object
      required: [id, title, author, sourceUrl, sourceStatus, hadiths]
      properties:
        id: { type: string }
        title: { type: string }
        author: { type: string }
        sourceUrl: { type: string }
        sourceStatus: { type: string, enum: [retrieved_pending_review, approved, unavailable] }
        hadiths:
          type: array
          items: { $ref: "#/components/schemas/Hadith" }
    Profile:
      type: object
      required: [id, name, role, onboarded]
      properties:
        id: { type: string }
        name: { type: string }
        role: { type: string, enum: [student, teacher] }
        onboarded: { type: boolean }
    ProfileInput:
      type: object
      required: [name, role]
      properties:
        name: { type: string, minLength: 2, maxLength: 100 }
        role: { type: string, enum: [student, teacher] }
    ProgressInput:
      type: object
      required: [currentHadith, completedIds, bookmarkedIds]
      properties:
        currentHadith: { type: integer, minimum: 1, maximum: 42 }
        completedIds:
          type: array
          maxItems: 42
          items: { type: integer, minimum: 1, maximum: 42 }
        bookmarkedIds:
          type: array
          maxItems: 42
          items: { type: integer, minimum: 1, maximum: 42 }
    StudyProgress:
      type: object
      required: [textId, currentHadith, completedIds, bookmarkedIds, updatedAt]
      properties:
        textId: { type: string }
        currentHadith: { type: integer }
        completedIds:
          type: array
          items: { type: integer }
        bookmarkedIds:
          type: array
          items: { type: integer }
        updatedAt: { type: string }
...
    SourceVersionInput:
      type: object
      additionalProperties: false
      required: [hadithNumber, text, printedPage, viewerPage, viewerUrl, edition, changeReason, rightsEvidence, rightsUrl]
      properties:
        hadithNumber: { type: integer, minimum: 1, maximum: 42 }
        text: { type: string, minLength: 10, maxLength: 12000 }
        printedPage: { type: integer, minimum: 1 }
        viewerPage: { type: integer, minimum: 1 }
        viewerUrl: { type: string, format: uri, maxLength: 1000 }
        edition: { type: string, minLength: 3, maxLength: 1000 }
        changeReason: { type: string, minLength: 5, maxLength: 2000 }
        rightsEvidence: { type: string, maxLength: 4000 }
        rightsUrl: { type: string, maxLength: 1000 }
    SourceVersion:
      allOf:
        - { $ref: "#/components/schemas/SourceVersionInput" }
        - type: object
          required: [id, version, status, scientificStatus, rightsStatus, createdAt, createdBy, history]
          properties:
            id: { type: string }
            version: { type: integer }
            status: { type: string, enum: [pending_review, approved, rejected, withdrawn] }
            scientificStatus: { type: string, enum: [pending, approved, rejected] }
            rightsStatus: { type: string, enum: [pending, cleared, rejected] }
            createdAt: { type: string }
            createdBy: { type: string }
            history: { type: array, items: { $ref: "#/components/schemas/ReviewAudit" } }
    SourceDecisionInput:
      type: object
      additionalProperties: false
      required: [decision, scientificStatus, rightsStatus, reason]
      properties:
        decision: { type: string, enum: [approved, rejected, withdrawn] }
        scientificStatus: { type: string, enum: [approved, rejected, pending] }
        rightsStatus: { type: string, enum: [cleared, rejected, pending] }
        reason: { type: string, minLength: 5, maxLength: 2000 }
    ReviewAudit:
      type: object
      required: [id, actorId, action, targetId, reason, createdAt]
      properties:
        id: { type: string }
        actorId: { type: string }
        action: { type: string }
        targetId: { type: string }
        reason: { type: string }
        createdAt: { type: string }
...