# lib/api-zod/src/generated/types/hadith.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HadithReviewStatus } from './hadithReviewStatus';

export interface Hadith {
  id: number;
  number: number;
  title: string;
  text: string;
  sourceUrl: string;
  sourcePage: number;
  sourceVersionId?: string;
  reviewStatus?: HadithReviewStatus;
  viewerPage?: number;
}