# lib/api-zod/src/generated/types/sourceDecisionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SourceDecisionInputDecision } from './sourceDecisionInputDecision';
import type { SourceDecisionInputRightsStatus } from './sourceDecisionInputRightsStatus';
import type { SourceDecisionInputScientificStatus } from './sourceDecisionInputScientificStatus';

export interface SourceDecisionInput {
  decision: SourceDecisionInputDecision;
  scientificStatus: SourceDecisionInputScientificStatus;
  rightsStatus: SourceDecisionInputRightsStatus;
  /**
     * @minLength 5
     * @maxLength 2000
     */
  reason: string;
}